Deck and Shuffler
So I haven't tested this out. It should work, but I'm not 100% sure it is debugged. Here is just the code for reference.

Changes: the vector is called deck_of_cards, there are 4 functions, 1.) void create_deck 2.)void shuffle_deck 3.)Value value_setter(j) 4.)Suit suit_setter(i).  The two setters explain themselves, create deck creates a deck of cards in order, and shuffle deck creates an array of random numbers 0-51, (none should repeat ), and then it swaps the indices of the vector and the array it is stored in